rename: SearchIcon.jsx SignUpIcon.jsx @/assets/icon으로 이동
