+++ src/src/components/common/SearchBar.jsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import { PropTypes } from "prop-types";
 import { useNavigate } from "react-router-dom";
 
-import SearchIcon from "@/assets/SearchIcon.jsx";
+import SearchIcon from "@/assets/icon/SearchIcon.jsx";
 import Button from "@/components/common/button/Button.jsx";
 import BUTTON_TYPE from "@/constants/BUTTON_TYPE.js";
 import routes from "@/constants/routes.js";
